+++ super-mario-3-se102/Hitbox.h
@@ -6,14 +6,14 @@ public:
 	/// <summary>
 	/// relativePosition: position relative to the position of entity
 	/// </summary>
-	Hitbox(const Utils::Vector2<float>& relativePosition, const Utils::Dimension<float>& dimension);
+	Hitbox(const Utils::Vector2<float>& relativePosition, const Utils::Dimension<int>& dimension);
 	Hitbox();
 
 	const Utils::Vector2<float>& GetRelativePosition() const;
-	const Utils::Dimension<float>& GetDimension() const;
+	const Utils::Dimension<int>& GetDimension() const;
 private:
 	Utils::Vector2<float> relativePosition;
-	Utils::Dimension<float> dimension;
+	Utils::Dimension<int> dimension;
 };
 typedef Hitbox* LPHitbox;
 typedef const Hitbox* LPConstHitbox;
